Move test out of lib.rs
