+++ tests/mod.rs
@@ -50,3 +50,34 @@ fn parse_extensions_v3_full() {
         Err(e) => eprintln!("{:?}", e),
     }
 }
+
+#[test]
+fn volumes() {
+    use docker_compose_types::Volumes;
+    use serde::Deserialize;
+
+    let v = r#"
+volumes:
+  - source: /host/path
+    target: /container/path
+    type: bind
+    read_only: true
+  - source: foobar
+    type: volume
+    target: /container/volumepath
+  - type: volume
+    target: /anonymous
+  - type: volume
+    source: foobar
+    target: /container/volumepath2
+    volume:
+      nocopy: true
+"#;
+
+    #[derive(Deserialize)]
+    #[allow(dead_code)]
+    struct Container {
+        volumes: Volumes,
+    }
+    let _parsed: Container = serde_yaml::from_str(v).unwrap();
+}
